feat(structure): add response structs for model chat and model text
This commit adds the response structs for the ModelChat and ModelText
structs in the gochatgptsdk package. These response structs include fields
such as ID, Object, Created, Choices, and Usage.

The response structs are necessary to properly handle the response data
returned by the OpenAI API when making requests using the ModelChat and
ModelText structs.

The following response structs have been added:
- ModelChatResponse
- ModelTextResponse

These response structs provide a structured representation of the response
data returned by the OpenAI API for chat and text models.

For more details on the response structs, please refer to the newly added
struct_response.go file.
